+++ runtime/src/commonMain/kotlin/dev/zacsweers/metro/Provider.kt
@@ -2,17 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro
 
+import dev.zacsweers.metro.internal.InstanceFactory
+
 /** A simple class that produces instances of [T]. */
 public fun interface Provider<T : Any> {
   public operator fun invoke(): T
 }
 
 /** A helper function to create a new [Provider] wrapper around a given [provider] lambda. */
-public inline fun <T : Any> provider(crossinline provider: () -> T): Provider<T> {
-  return Provider { provider() }
+public inline fun <T : Any> provider(crossinline provider: () -> T): Provider<T> = Provider {
+  provider()
 }
 
-/** Creates a [Provider] wrapper around the given [value]. */
-public fun <T : Any> providerOf(value: T): Provider<T> {
-  return Provider { value }
-}
+/** Returns a [Provider] wrapper around the given [value]. */
+public fun <T : Any> providerOf(value: T): Provider<T> = InstanceFactory(value)