+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 - **Fix**: Fix a bug where `Map<Key, Provider<Value>>` multibindings weren't always unwrapped correctly.
 - **Fix**: Fix `Map<Key, Provider<Value>>` type keys not correctly interpreting the underlying type key as `Map<Key, Value>`.
 - **Change**: Change `InstanceFactory` to a value class.
+- **Change**: Make `providerOf` use `InstanceFactory` under the hood.
 
 0.1.1
 -----